+++ CHANGELOG.md
@@ -2,6 +2,9 @@
   * Fatal crash reports will now be written in sync mode to shared preferences 
   * Added explicit storage mode where the event queue and request queue are not writen to persistent storage until a signal is given
 
+  * Removed the deprecated "disableLocation" method
+  * Removed the deprecated "setLocation" method
+
 ## 22.02.3
   * Added CountlyConfigPush object which is used during CountlyPush init.
   * Added a way to add allowed package names for push notification intent security.